+++ src/components/Profile.tsx
@@ -1,13 +1,17 @@
 import {useAtomValue} from 'jotai';
 import { authAtom } from '../../store/atoms/';
+import {Button} from '@mui/material'
 
 const Profile= () => {
-    const {username, email} = useAtomValue(authAtom).user
+    const {username, email, isPro} = useAtomValue(authAtom).user
     return (
     <div>
         <h1> Profile</h1>   
         <p>Username: {username || "N/A"}</p> 
         <p>Email: {email}</p> 
+
+        <Button>{isPro ? "Downgrade Account" : "Upgrade to Pro"}</Button>
+
     </div>
     );
 }