+++ src/components/StockList.tsx
@@ -1,18 +1,25 @@
 
 import {useAtomValue, useSetAtom} from 'jotai';
+import {useEffect} from 'react';
 import { paramsAtom, stocksAtom, updateAllocation } from '../../store/atoms/';
 import { StockData, StockAllocation, IsChanged } from '../../api';
-import {Box, Paper, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Slider } from '@mui/material';
+import {Box, Paper, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Slider, CircularProgress, Button } from '@mui/material';
 import { useState } from 'react';
 import SaveIcon from '@mui/icons-material/Save';
+import { Link } from 'react-router-dom';
 
 const StockList = () => {
     const [adjustedAllocation, setAdjustedAllocation] = useState<StockAllocation>({});
     const [isChanged, setIsChanged] = useState<IsChanged>({});
     const stockData: StockData[] = useAtomValue(stocksAtom).stockData
+    const loading: boolean = useAtomValue(stocksAtom).loading
     const totalCapital = useAtomValue(paramsAtom).capital
     const allocationData: StockAllocation = useAtomValue(stocksAtom).allocationData
     const updateAlloc = useSetAtom(updateAllocation)
+
+    useEffect(()=>{
+        console.log("AHHH: ", stockData)
+    }, [stockData])
 
     const handleAllocationAdjustment = (symbol: string, value: string) =>{
         let adjAlloc = {...adjustedAllocation}
@@ -29,26 +36,36 @@ const StockList = () => {
         vals.forEach(val=>total+=val.percent)
         updateAlloc(adjustedAllocation, total)
         setIsChanged({})
-        
     }
   
     return (
         <div>
-            {stockData.length > 0 
+            {
+                loading ?  <Box sx={{ display: 'flex', justifyContent: 'center', padding: 20 }}>
+                            <CircularProgress />
+                            </Box>
+                            : 
+                <>
+                { stockData.length > 0 
             ? <TableContainer component={Paper}>
             <Table sx={{ minWidth: 650 }} size="small" aria-label="a dense table">
                 <TableHead>
                     <TableRow>
                         <TableCell align="left">Symbol</TableCell>
                         <TableCell>Allocation (%)</TableCell>
-                        <TableCell>Capital Allocation</TableCell>
-                        <TableCell>One Month Return</TableCell>
-                        <TableCell>Three Month Return</TableCell>
+                        <TableCell>Allocation ($)</TableCell>
+                        <TableCell>1 Month ROI</TableCell>
+                        <TableCell>3 Month ROI</TableCell>
+                        <TableCell>6 Month ROI</TableCell>
+                        <TableCell>1 Year ROI</TableCell>
+                        <TableCell>5 Year ROI</TableCell>
+                        {/* <TableCell>All Time Return</TableCell> */}
                         <TableCell align="right">Actions</TableCell>
                     </TableRow>
                 </TableHead>
                 <TableBody>
                 {stockData.map((stock) => {
+                    let {oneMonthRoi, threeMonthRoi, sixMonthRoi, oneYearRoi, fiveYearRoi} = stock
                     return <TableRow
                         key={stock.symbol}
                         sx={{ '&:last-child td, &:last-child th': { border: 0 } }}
@@ -68,10 +85,19 @@ const StockList = () => {
                             <p>${adjustedAllocation[stock.symbol]?.cap || allocationData[stock.symbol].cap}</p>
                         </TableCell>
                         <TableCell>
-                            <p className={stock.oneMonthRoi && stock.oneMonthRoi > 0 ? 'positive' : 'negative'}>{stock.oneMonthRoi && stock.oneMonthRoi > 0 ? "+" : null}{stock.oneMonthRoi ? Math.round((stock.oneMonthRoi * 100) * 100) / 100 + "%" : "N/A"}</p>
+                            <p className={oneMonthRoi && oneMonthRoi > 0 ? 'positive' : 'negative'}>{oneMonthRoi && oneMonthRoi > 0 ? "+" : null}{oneMonthRoi ? Math.round((oneMonthRoi * 100) * 100) / 100 + "%" : "N/A"}</p>
                         </TableCell>
                         <TableCell>
-                            <p className={stock.threeMonthRoi && stock.threeMonthRoi > 0 ? 'positive' : 'negative'}>{stock.threeMonthRoi && stock.threeMonthRoi > 0 ? "+" : null}{stock.threeMonthRoi ?  Math.round((stock.threeMonthRoi * 100) * 100) / 100 + "%" : "N/A"}</p>
+                            <p className={threeMonthRoi && threeMonthRoi > 0 ? 'positive' : 'negative'}>{threeMonthRoi && threeMonthRoi > 0 ? "+" : null}{threeMonthRoi ?  Math.round((threeMonthRoi * 100) * 100) / 100 + "%" : "N/A"}</p>
+                        </TableCell>
+                        <TableCell>
+                            {sixMonthRoi && sixMonthRoi !== "Upgrade" ?  <p className={sixMonthRoi && sixMonthRoi > 0 ? 'positive' : 'negative'}>{sixMonthRoi > 0 ? "+" : null}{sixMonthRoi ?  Math.round((sixMonthRoi * 100) * 100) / 100 + "%" : "N/A"}</p> : <Link className="upgradeLink" to="/profile">Unlock With Pro</Link>}
+                        </TableCell>
+                        <TableCell>
+                            {oneYearRoi && oneYearRoi !== "Upgrade" ?  <p className={oneYearRoi > 0 ? 'positive' : 'negative'}>{oneYearRoi > 0 ? "+" : null}{sixMonthRoi ?  Math.round((oneYearRoi * 100) * 100) / 100 + "%" : "N/A"}</p> : <Link className="upgradeLink" to="/profile">Unlock With Pro</Link> }
+                        </TableCell>
+                        <TableCell>
+                        {fiveYearRoi && fiveYearRoi !== "Upgrade" ?  <p className={fiveYearRoi > 0 ? 'positive' : 'negative'}>{fiveYearRoi > 0 ? "+" : null}{fiveYearRoi ?  Math.round((fiveYearRoi * 100) * 100) / 100 + "%" : "N/A"}</p> : <Link className="upgradeLink" to="/profile">Unlock With Pro</Link> }
                         </TableCell>
                         <TableCell align="right">
                             <Box sx={{ display: 'flex', justifyContent: 'center', mt: 2 }}>
@@ -85,6 +111,8 @@ const StockList = () => {
             </Table>
         </TableContainer> 
             : <p>Select some tickers to analyze!</p>
+            }
+                </>
             }
         </div>
     )