+++ src/components/auth/ProtectedRoute.tsx
@@ -1,4 +1,4 @@
-import { Navigate } from 'react-router-dom';
+import { Navigate, useLocation } from 'react-router-dom';
 import { useAuth } from './AuthContext.tsx';
 
 interface ProtectedRouteProps {
@@ -7,6 +7,12 @@ interface ProtectedRouteProps {
 
 export const ProtectedRoute = ({ children }: ProtectedRouteProps) => {
   const { isAuthenticated } = useAuth();
+  const location = useLocation();
+
+   // Allow access to /forgot-password even if not authenticated
+  if (!isAuthenticated && location.pathname === "/forgot-password") {
+    return children;
+  }
 
   return isAuthenticated ? children : <Navigate to="/login" />;
 };